Extracted chat components to separate files
