+++ src/components/chat/index.tsx
@@ -1,0 +1,5 @@
+import { ChatHeader } from "./chatHeader";
+import { ChatMessage } from "./chatMessage";
+import { ChatComponent } from "./chatComponent";
+
+export { ChatHeader, ChatMessage, ChatComponent };
